Create Website Title and Main Tabs
Declared the title of the website to be MReviews, and made a header. Made a heading1 called MReviews that is hyperlinked to the index.html file. Also created an unordered list for the 4 main tabs: Home, All Movies, Genre, and Forum. The Home item is also hyperlinked to the index.html file.